+++ src/BasicInputs/SelectInput.tsx
@@ -14,7 +14,7 @@ export const SelectInput = forwardRef<
   ElementRef<typeof Select>,
   PropsWithoutRef<FormControlChildProps> &
     ComponentPropsWithoutRef<typeof Select>
->(({ disabled, className, ...props }, ref) => {
+>(({ disabled, className, selectoptions, options, ...props }, ref) => {
   const defaultStyling: ComponentProps<typeof Select>['classNames'] = {
     clearIndicator: ({ isFocused }) =>
       cn(
@@ -104,7 +104,7 @@ export const SelectInput = forwardRef<
   return (
     <div
       className={cn(
-        'rounded-md focus-within:outline focus-within:outline-2 focus-within:outline-offset-2',
+        'rounded-md ring-offset-background focus-within:outline-none focus-within:ring-2 focus-within:ring-ring focus-within:ring-offset-2',
         disabled && 'cursor-not-allowed',
         className,
       )}
@@ -113,8 +113,8 @@ export const SelectInput = forwardRef<
         ref={ref}
         unstyled
         classNames={defaultStyling}
-        menuPortalTarget={document.body}
         isDisabled={disabled}
+        options={options ?? selectoptions}
         {...props}
       />
     </div>