+++ src/ModelForm/ModelFormField.tsx
@@ -39,42 +39,51 @@ export const ModelFormField = ({
   const valueOptions = useModelFormStore(
     (state) => state.fieldOptions?.[field]?.valueOptions ?? [],
   );
+  const inputProps = useModelFormStore(
+    (state) => state.fieldOptions?.[field]?.inputProps,
+  );
+  const displayProps = useModelFormStore(
+    (state) => state.fieldOptions?.[field]?.displayProps,
+  );
   const DisplayComponent = useConveyorStore(
     (state) => state.typeOptions?.[type]?.DisplayComponent ?? (() => null),
   );
   const InputComponent = useConveyorStore(
     (state) => state.typeOptions?.[type]?.InputComponent ?? (() => null),
   );
+
   return (
     <Slot slotKey={field}>
       <div className={cn('flex flex-col space-y-2', className)} {...divProps}>
         {children === undefined ? (
           <>
-            <FormLabel name={field} required={required} />
+            <FormLabel name={field} required={required}>
+              {label}
+            </FormLabel>
             {editable ? (
               <>
                 <Lens lens={DataLens.DISPLAY}>
                   <FormDisplay name={field}>
-                    <DisplayComponent />
+                    <DisplayComponent {...displayProps} />
                   </FormDisplay>
                 </Lens>
                 <Lens lens={DataLens.INPUT}>
                   <FormControl
                     name={field}
-                    options={valueOptions}
+                    selectoptions={valueOptions}
                     rules={{
                       required: required && `${label} is required.`,
                       ...rules,
                     }}
                   >
-                    <InputComponent />
+                    <InputComponent {...inputProps} />
                   </FormControl>
                   <FormError name={field} />
                 </Lens>
               </>
             ) : (
               <FormDisplay name={field}>
-                <DisplayComponent />
+                <DisplayComponent {...displayProps} />
               </FormDisplay>
             )}
           </>