+++ src/ModelTable/ModelTableCell.tsx
@@ -41,6 +41,12 @@ export const ModelTableCell = ({
   const valueOptions = useModelTableStore(
     (state) => state.columnOptions?.[field]?.valueOptions ?? [],
   );
+  const inputProps = useModelTableStore(
+    (state) => state.columnOptions?.[field]?.inputProps,
+  );
+  const displayProps = useModelTableStore(
+    (state) => state.columnOptions?.[field]?.displayProps,
+  );
   const DisplayComponent = useConveyorStore(
     (state) => state.typeOptions?.[type]?.DisplayComponent ?? (() => null),
   );
@@ -81,25 +87,25 @@ export const ModelTableCell = ({
             <>
               <Lens lens={DataLens.DISPLAY}>
                 <FormDisplay name={field}>
-                  <DisplayComponent />
+                  <DisplayComponent {...displayProps} />
                 </FormDisplay>
               </Lens>
               <Lens lens={DataLens.INPUT}>
                 <FormControl
                   name={field}
-                  options={valueOptions}
+                  selectoptions={valueOptions}
                   rules={{
                     required: required && `${label} is required.`,
                     ...rules,
                   }}
                 >
-                  <InputComponent />
+                  <InputComponent {...inputProps} />
                 </FormControl>
               </Lens>
             </>
           ) : (
             <FormDisplay name={field}>
-              <DisplayComponent />
+              <DisplayComponent {...displayProps} />
             </FormDisplay>
           )
         ) : (